move attribute validation code to separate validation class + decorator
