+++ molgenis-data-platform/src/main/java/org/molgenis/data/platform/decorators/MolgenisRepositoryDecoratorFactory.java
@@ -33,6 +33,8 @@ import org.molgenis.data.transaction.TransactionInformation;
 import org.molgenis.data.validation.EntityAttributesValidator;
 import org.molgenis.data.validation.ExpressionValidator;
 import org.molgenis.data.validation.RepositoryValidationDecorator;
+import org.molgenis.data.validation.meta.AttributeRepositoryValidationDecorator;
+import org.molgenis.data.validation.meta.AttributeValidator;
 import org.molgenis.data.validation.meta.EntityTypeRepositoryValidationDecorator;
 import org.molgenis.data.validation.meta.EntityTypeValidator;
 import org.molgenis.security.core.MolgenisPermissionService;
@@ -74,6 +76,7 @@ public class MolgenisRepositoryDecoratorFactory implements RepositoryDecoratorFa
 	private final L3Cache l3Cache;
 	private final LanguageService languageService;
 	private final EntityTypeDependencyResolver entityTypeDependencyResolver;
+	private final AttributeValidator attributeValidator;
 
 	@Autowired
 	public MolgenisRepositoryDecoratorFactory(EntityManager entityManager,
@@ -84,7 +87,8 @@ public class MolgenisRepositoryDecoratorFactory implements RepositoryDecoratorFa
 			PasswordEncoder passwordEncoder, L1Cache l1Cache, L2Cache l2Cache,
 			TransactionInformation transactionInformation, EntityListenersService entityListenersService,
 			MolgenisPermissionService permissionService, EntityTypeValidator entityTypeValidator, L3Cache l3Cache,
-			LanguageService languageService, EntityTypeDependencyResolver entityTypeDependencyResolver)
+			LanguageService languageService, EntityTypeDependencyResolver entityTypeDependencyResolver,
+			AttributeValidator attributeValidator)
 
 	{
 		this.entityManager = requireNonNull(entityManager);
@@ -107,6 +111,7 @@ public class MolgenisRepositoryDecoratorFactory implements RepositoryDecoratorFa
 		this.l3Cache = requireNonNull(l3Cache);
 		this.languageService = requireNonNull(languageService);
 		this.entityTypeDependencyResolver = requireNonNull(entityTypeDependencyResolver);
+		this.attributeValidator = requireNonNull(attributeValidator);
 	}
 
 	@Override
@@ -169,6 +174,9 @@ public class MolgenisRepositoryDecoratorFactory implements RepositoryDecoratorFa
 		}
 		else if (repo.getName().equals(ATTRIBUTE_META_DATA))
 		{
+			repo = (Repository<Entity>) (Repository<? extends Entity>) new AttributeRepositoryValidationDecorator(
+					(Repository<Attribute>) (Repository<? extends Entity>) repo, attributeValidator);
+
 			repo = (Repository<Entity>) (Repository<? extends Entity>) new AttributeRepositoryDecorator(
 					(Repository<Attribute>) (Repository<? extends Entity>) repo, systemEntityTypeRegistry, dataService,
 					permissionService);
