+++ molgenis-data-validation/src/main/java/org/molgenis/data/validation/meta/AttributeRepositoryValidationDecorator.java
@@ -1,0 +1,62 @@
+package org.molgenis.data.validation.meta;
+
+import org.molgenis.data.AbstractRepositoryDecorator;
+import org.molgenis.data.Repository;
+import org.molgenis.data.meta.model.Attribute;
+
+import java.util.stream.Stream;
+
+import static java.util.Objects.requireNonNull;
+
+public class AttributeRepositoryValidationDecorator extends AbstractRepositoryDecorator<Attribute>
+{
+	private final Repository<Attribute> decoratedRepo;
+	private final AttributeValidator attributeValidator;
+
+	public AttributeRepositoryValidationDecorator(Repository<Attribute> decoratedRepo,
+			AttributeValidator attributeValidator)
+	{
+		this.decoratedRepo = requireNonNull(decoratedRepo);
+		this.attributeValidator = requireNonNull(attributeValidator);
+	}
+
+	@Override
+	protected Repository<Attribute> delegate()
+	{
+		return decoratedRepo;
+	}
+
+	@Override
+	public void update(Attribute attr)
+	{
+		attributeValidator.validateUpdate(attr);
+	}
+
+	@Override
+	public void update(Stream<Attribute> attrs)
+	{
+		decoratedRepo.update(attrs.filter(attr ->
+		{
+			Attribute currentAttr = findOneById(attr.getIdentifier());
+			attributeValidator.validateUpdate(attr);
+			return true;
+		}));
+	}
+
+	@Override
+	public void add(Attribute attr)
+	{
+		attributeValidator.validateAdd(attr);
+		decoratedRepo.add(attr);
+	}
+
+	@Override
+	public Integer add(Stream<Attribute> attrs)
+	{
+		return decoratedRepo.add(attrs.filter(attr ->
+		{
+			attributeValidator.validateAdd(attr);
+			return true;
+		}));
+	}
+}
